+++ src/geo/projection/globe_transform.ts
@@ -422,6 +422,10 @@ export class GlobeTransform implements ITransform {
         return;
     }
 
+    locationToScreenPointAtElevation(lnglat: LngLat, elevation: number): Point {
+        return this.currentTransform.locationToScreenPointAtElevation(lnglat, elevation);
+    }
+
     locationToScreenPoint(lnglat: LngLat, terrain?: Terrain): Point {
         return this.currentTransform.locationToScreenPoint(lnglat, terrain);
     }